Move EMPTY_ANALYZED_CONTRACT to testUtils.ts
