+++ packages/discovery/src/discovery/analysis/AddressAnalyzer.ts
@@ -52,28 +52,6 @@ export interface AnalyzedContract {
   usedTypes?: DiscoveryCustomType[]
 }
 
-export const EMPTY_ANALYZED_CONTRACT: AnalyzedContract = {
-  type: 'Contract',
-  address: EthereumAddress.ZERO,
-  name: '',
-  deploymentTimestamp: UnixTime.ZERO,
-  deploymentBlockNumber: 0,
-  derivedName: undefined,
-  isVerified: false,
-  proxyType: '',
-  implementations: [],
-  values: {},
-  fieldsMeta: {},
-  errors: {},
-  abis: {},
-  sourceBundles: [],
-  extendedTemplate: undefined,
-  ignoreInWatchMode: undefined,
-  relatives: {},
-  selfMeta: undefined,
-  usedTypes: [],
-}
-
 export interface ExtendedTemplate {
   template: string
   reason: 'byExtends' | 'byReferrer' | 'byShapeMatch'
