+++ packages/discovery/src/discovery/output/saveDiscoveryResult.test.ts
@@ -2,10 +2,8 @@ import { EthereumAddress, UnixTime } from '@l2beat/shared-pure'
 import { expect } from 'earl'
 import { map } from 'lodash'
 
-import {
-  AnalyzedContract,
-  EMPTY_ANALYZED_CONTRACT,
-} from '../analysis/AddressAnalyzer'
+import { AnalyzedContract } from '../analysis/AddressAnalyzer'
+import { EMPTY_ANALYZED_CONTRACT } from '../utils/testUtils'
 import { getSourceOutputPath } from './saveDiscoveryResult'
 
 describe(getSourceOutputPath.name, () => {
